Updated token delegators
