+++ VideoApi/VideoApi.Services/Handlers/Kinly/KinlyApiTokenDelegatingHandler.cs
@@ -1,0 +1,9 @@
+﻿using VideoApi.Common.Security.Supplier.Kinly;
+using VideoApi.Services.Handlers.Base;
+
+namespace VideoApi.Services.Handlers.Kinly;
+
+public class KinlyApiTokenDelegatingHandler : CustomApiTokenDelegatingHandler
+{
+    public KinlyApiTokenDelegatingHandler(IKinlyJwtProvider tokenProvider) : base(tokenProvider){}
+}
