+++ VideoApi/VideoApi.Services/Handlers/Vodafone/VodafoneApiTokenDelegatingHandler.cs
@@ -1,0 +1,9 @@
+﻿using VideoApi.Common.Security.Supplier.Vodafone;
+using VideoApi.Services.Handlers.Base;
+
+namespace VideoApi.Services.Handlers.Vodafone;
+
+public class VodafoneApiTokenDelegatingHandler : CustomApiTokenDelegatingHandler
+{
+    public VodafoneApiTokenDelegatingHandler(IVodafoneJwtProvider tokenProvider) : base(tokenProvider){}
+}
